bluethooth

+++ berceau/Firebase.h
@@ -1,5 +1,5 @@
-int getLedFirebase();
-void setLedFirebase(int value);
+uint8_t getLedFirebase();
+void setLedFirebase(uint8_t value);
 void initFirebase();
 String readSsid();
 String readPassword();
